refactor: cleanup unused python logic files
- Remove backup/duplicate service files in dominial/services/
- Merge logic from patch_hierarquia_arvore_service.py into main service
- Remove obsolete management commands (tests, one-off patches)
- Remove duplicate documento_service_consolidado.py
