+++ .commit_msg_alteracao_ti.txt
@@ -1,0 +1,22 @@
+feat(admin): Adiciona ferramenta para correção de TI de imóveis no painel admin
+
+- Implementa ImovelAdmin customizado com funcionalidade de alteração de TI
+- Adiciona view customizada para confirmação e alteração de TI com validações
+- Exibe informações sobre documentos e lançamentos relacionados
+- Registra alterações no campo observações do imóvel para auditoria
+- Adiciona templates para página de alteração e botão na edição
+- Inclui documentação completa da estratégia e guia de testes
+
+Funcionalidades:
+- Listagem melhorada com coluna de documentos/lançamentos
+- Validações de segurança (não permite selecionar TI atual)
+- Avisos visuais para imóveis com dados relacionados
+- Registro completo de auditoria com data, usuário e motivo
+- Interface intuitiva com confirmações duplas
+
+Arquivos:
+- dominial/admin.py: ImovelAdmin customizado
+- templates/admin/dominial/imovel/alterar_ti.html: Página de alteração
+- templates/admin/dominial/imovel/change_form.html: Botão na edição
+- docs/ESTRATEGIA_ALTERACAO_TI_ADMIN.md: Documentação técnica
+- docs/GUIA_TESTE_ALTERACAO_TI.md: Guia de testes
