+++ COMANDOS_DEPLOY_DOCKER.txt
@@ -1,0 +1,32 @@
+# 🚀 Comandos Deploy - Docker Compose (Copy & Paste)
+
+## Ordem Correta:
+
+# 1. Backup
+docker-compose exec db pg_dump -U $DB_USER $DB_NAME > backup_antes_migracao_matricula_$(date +%Y%m%d_%H%M%S).sql
+ls -lh backup_antes_migracao_matricula_*.sql
+
+# 2. Atualizar código
+git pull origin main
+
+# 3. Reconstruir container (IMPORTANTE!)
+docker-compose up -d --build web
+
+# 4. Verificar dados (AGORA o comando existe)
+docker-compose exec web python manage.py verificar_matricula_constraint
+
+# 5. Aplicar migração
+docker-compose exec web python manage.py migrate
+
+# 6. Verificar migração
+docker-compose exec web python manage.py showmigrations dominial | grep 0042
+
+# 7. Verificar constraints
+docker-compose exec db psql -U $DB_USER -d $DB_NAME -c "\d dominial_imovel" | grep -E "(unique|index)"
+
+# 8. Reiniciar
+docker-compose restart web
+
+# 9. Verificar logs
+docker-compose logs web --tail=20
+
