+++ wrangler.toml
@@ -1,0 +1,56 @@
+# Cloudflare Workers/Pages Configuration for Cadeia Dominial
+# Generated for Django 5.2 application deployment
+
+name = "cadeia-dominial"
+compatibility_date = "2025-01-07"
+compatibility_flags = ["nodejs_compat"]
+
+# Workers deployment
+[build]
+command = "pip install -r requirements.txt"
+
+[vars]
+DJANGO_SETTINGS_MODULE = "cadeia_dominial.settings_prod"
+DEBUG = "false"
+SECRET_KEY = "$DJANGO_SECRET_KEY"
+
+# Database configuration (using external PostgreSQL)
+# Note: For Cloudflare D1, use [[d1_databases]] below
+[env.production.vars]
+DATABASE_URL = "$DATABASE_URL"
+
+# D1 Database (if using Cloudflare D1 instead of external PostgreSQL)
+# [[d1_databases]]
+# binding = "DB"
+# database_name = "cadeia_dominial_db"
+# database_id = "your-d1-database-id"
+
+# KV Namespaces for caching/static files
+[[kv_namespaces]]
+binding = "CACHE"
+id = "$CACHE_KV_ID"
+
+[[kv_namespaces]]
+binding = "STATIC_FILES"
+id = "$STATIC_KV_ID"
+preview_id = "$STATIC_KV_PREVIEW_ID"
+
+# R2 Bucket for file uploads (matrículas, documents)
+[[r2_buckets]]
+binding = "R2_BUCKET"
+bucket_name = "cadeia-dominial-uploads"
+
+# Routes configuration
+[[routes]]
+pattern = "/*"
+zone_name = "cadeiadominial.com.br"
+
+# Custom build options for Django static files
+[build.filters]
+include = ["**/*.py", "**/templates/**", "static/**"]
+exclude = ["venv/**", ".git/**", "node_modules/**"]
+
+# Dev configuration for local development
+[dev]
+port = 8000
+local = true
